Move user profile view tests to dedicate file
Part of test suite reorganization
